refactor: use resolver
